Another refactor pass. Initial menu system in place.
Commits will start to be more frequent.
Built the credential class to handle secure strings or profiles being sent to it.
Built the ECCInstanceFinder to take the credential class and be a utility to the toolkit.
Built the userinterface out to actually take in and process data
New Utilities class to handle various functionalities, like secure line reads
Added in the AWS libraries as dependencies